requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.19.4
pycodestyle==2.8.0
requests==2.27.1
SQLAlchemy==1.4.32
toml==0.10.2
urllib3==1.26.9
Werkzeug==2.0.3
